+++ base/src/systems/unable_to_locate_account.rs
@@ -1,9 +1,7 @@
 use shared::prelude::*;
 
-pub fn unable_to_locate_account(
+pub fn unable_to_locate_account_move_user(
     mut unable_to_locate_account_rx: EventReader<UnableToLocateAccountEvent>,
-    mut text_event_tx: EventWriter<TextEvent>,
-    mut show_prompt_rx: EventWriter<ShowPromptEvent>,
     mut query: Query<&mut UserSessionData>,
 ) {
     for ev in unable_to_locate_account_rx.read() {
@@ -15,11 +13,5 @@ pub fn unable_to_locate_account(
             }
         };
         user_session_data.status = UserStatus::NeedUsername;
-
-        text_event_tx.send(TextEvent::from_str(
-            ev.0,
-            "Those credentials don't seem to match. Try again.",
-        ));
-        show_prompt_rx.send(ShowPromptEvent(ev.0));
     }
 }
